Added cast to float

+++ baselines/pois/pois.py
@@ -424,7 +424,7 @@ def learn(make_env, make_policy, *,
         print("Interval size:", interval_size)
         # Clip to avoid overflow and cluster
         ep_return = tf.clip_by_value(ep_return, rew_min, rew_max)
-        ep_return = tf.floordiv(ep_return, interval_size) * interval_size
+        ep_return = tf.cast(tf.floordiv(ep_return, interval_size) * interval_size, tf.float32)
         tf.add_to_collection('prints', tf.Print(ep_return, [ep_return], 'ep_return_clustered', summarize=20))
     else:
         raise Exception('Unrecognized reward clustering scheme.')
